requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
coverage==7.3.2
Django==4.2.5
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
packaging==23.2
Pillow==10.0.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
